Supplementary scripts and corrections
A whole bunch of commits here, some of which should have been better
separated, but lots of this was done during the Sept 2016 Nouméa trip,
which was very rushed.

Created two post-sim analysis scripts to calculate particle densities
and compare multiple particle density outputs for the proto-tagging
simulator that was thrown together in Nouméa.

This commit also contains several improvements to the creation of
SEAPODYM diffusivity fields, along with a test script that generates
and loads diffusivity files for comparison.
This follows on continuing improvements/corrections sent through from
Inna, and includes some corrections to the calculation of diffusivity,
and options to generate K fields in Nm/mo (the units used in SEAPODYM)