Fix for expanded memory footprint
Somewhat of a hack to fix the slowly increasing memory footprint of
grid objects  on Raijin, due to arbitrary use of 30 day months.

Increased month loop to be 31 days each iteration. This slowly eats up
days of the following month through the course of the simulation, but
as we have three months in memory at once, this should be fine