Advance time bugfix
It seems that the advance time function was cutting off the same number
of days as the new loaded month, rather than the previously loaded
month. This has been fixed.

Also, a couple more minor additions to ensure things work with BRAN, as
well as hardcoding the FAD_duration grid constant to be greater than a
year