Runtime pathway land testing
Altered movement kernels to now test each movement for land mask
crossing at quarterly intervals of the total movement vector each
timestep. This should eliminate all “jumping” over landmasses, such as
Japan. Illegal moves result in a re-draw of the diffusion component of
the movement, up to 500 times. After which, the particle is
deactivated, for the moment.

Slight update to options for gradient calculation