Parameter file reading
Made additions that allow switching from increasingly long command line
arguments to reading a parameter file.
This includes linking to a SEAPODYM parfile to pull out the relevant
parameters for the appropriate run (now we are generalising for both
mercator and interim runs)
This includes multiple fisheries, which will be implemented next.

Generalised the use of bathymetry data for the landmasking. There must
now be a separate field for bathymetry data (for the SEAPODYM
replication it was previously hard coded to look for the third depth
layer in the physical forcing file)