Edits to fieldset advance time
Made some changes to the advance time method so that new field
snapshots are only loaded if there are files for them, otherwise the
grid is simply shortened to the newest month available. This way we can
keep running the simulation until the last available time.

Also, previously the length of the first month advection loop was
hardcoded to for April. Have now made this generic