Changes for latest parcels master
This commit makes some small but required changes to ensure running on the latest parcels master. These include moving the execution start time to the particle set construction, and a few changes to the gradient calculation function.

This should *hopefully* make things a lot more stable and negate the need to use old stale parcels branches, particularly for the gradient calculation stuff